Shop V3

+++ clicker/Assets/Ganto/Scripts/ShopItemSO.cs
@@ -8,5 +8,5 @@ public class ShopItemSO : ScriptableObject
 {
     public string title;
     public string description;
-    public int basecost;
+    public float basecost;
 }
